+++ store/searchFilterStore.ts
@@ -4,4 +4,6 @@ export const searchFilterStore = reactive({
   search: '',
   orderBy: 'asc',
   type: '',
+  results: 0,
+  total: 0,
 })
